+++ CHANGELOG.md
@@ -1,1 +1,5 @@
+# Indevelopment
 
+1.  Support json, ini format.
+2.  Reload config when the file changes.
+3.  Allow to hook change.